requirements-test.txt
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component
aioresponses
